Move validate.ReadConfig to new config.go file
